refactor data fetching
Fetch only the bits of the JSON we need into an internal "record"
format, and then act on that.

This is a significant speedup (~13s ~> ~8s) on my laptop for reasons
I'm not entirely clear on. I expect they relate to reusing the
simdjson parser and also somehow making it inflate less of the
JSON. It also isolates the memory unsafety to `fetch_all_records` and
sets up for the next commit.