fix: better sticky headers
Due to Chrome bug 702927 we can't sticky the thead, so instead we reuse
the code that existed but properly grab the right set of headers.

In Chrome (again...) this means the header borders are transparent (because
we're stickying the cells, not their borders, I guess), so I've also removed
the left/right borders in the thead cells to avoid a visual artifact. This
seems fine, but can be reverted if appropriate.