fix: use a td instead of a th for state names, sticky header
Closes #127 and fixes up the sticky headers to be cleaner: they now use
an expected rem height that matches the height of the first row, and
auto-set flag heights to fill that as well.

Appears to work in Chrome and FF, with a minor visual artifact in FF
when scrolling up into a new table for a couple rows (with a
single-pixel gap between the first two rows). Unclear what can be done
about that, if anything, so I'll leave it be.