Fix Notification typo, calculate interval
Change `Notifcation` to `Notification` to fix ReferenceError and make
notifications work

Change `setInterval(..., 5 * 60 * 1000)` to `setTimeout`, and set the timeout
to a calculated number of milliseconds so that the page is updated when the
repo actually changes, rather than 5 minutes after the page loaded. Fall back
to a 1 minute interval if the repo hasn't updated on time.