Move JavaScript to its own file
Problem: We can cache this JavaScript rather than loading it on every
HTTP request, which also make it slightly easier to test and edit (at
least for me).

Solution: Take the inline `<script>` tags and move them to `main.js`,
formatted with Prettier since it's going to be a huge diff anyway.