Fix batch breakdown calculations
Previously, the batch breakdown was calculated using the total new votes
(which relies on the assumption that all new votes were for the top two
candidates).

This was exacerbated when the underlying data source had a number of
third-party and write-in vote counts revised down (which caused negative
percentages, or percentages over 100%).