Use latest data for third-party estimate when calculating hurdle
To estimate the third-party votes across the entire vote population, we
should use the largest sample size we have, which is the latest data
we've received.

This means using the latest batch to calculate hurdles for older batches
and, consequently, hurdle values for older batches might change slightly
as new data is received.

In practice, this change makes a minute difference. The largest change
for the current data was a 0.276 percentage point difference in the
proportion of third-party votes, and a ~0.1 percentage point difference
in calculated hurdle.